requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
ddddocr==1.4.2
flatbuffers==2.0
h11==0.13.0
idna==3.3
loguru==0.6.0
numpy==1.22.2
onnxruntime==1.10.0
opencv-python==4.5.5.62
outcome==1.1.0
Pillow==9.0.1
protobuf==3.19.4
pycparser==2.21
pyOpenSSL==22.0.0
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
